Fix C4244 narrowing warnings via type promotion
Promote function signatures and local variables to wider types to eliminate
MSVC C4244 warnings without losing precision:

- dwgbuffer: numRemainingBytes() -> int64_t, moveBitPos(int64_t),
  crc8() start/end -> int64_t, getUCSStr/get16bitStr size -> uint32_t
- dwgreader: maxPos -> uint64_t, startPos -> uint64_t,
  dwgSectionInfo compressed/encrypted -> uint64_t, pages map key -> uint64_t
- dwgutil: compressedSize/decompSize/compressedPos/decompPos -> uint64_t,
  compressedByte/decompByte index -> uint64_t, decompress18 locals -> uint64_t,
  decrypt18Hdr loop vars -> uint64_t, twoByteOffset litCount param -> uint64_t
- drw_entities/drw_objects: ms -> static_cast<uint32_t>(buf->size()),
  numRemainingBytes local -> int64_t
- dxfwriter: writeInt64 byte extraction -> static_cast<char>
- libdwgr/dwgreader15/18/21: static_cast<size_t> at allocator boundaries

Use static_cast only where narrowing is intentional per DWG spec
(e.g., 32-bit XOR mask, allocator boundaries on 32-bit platforms).

(cherry picked from commit 7f8b809e6a69a82f655c27efa855d131e98b4d6f)

+++ src/intern/dwgreader.h
@@ -111,9 +111,9 @@ public:
     dwgSectionInfo()=default;
     std::int32_t Id{-1}; //section Id, 2000-   rd15 rd18
     std::string name; //section name rd18
-    std::uint32_t compressed{1};//is compressed? 1=no, 2=yes rd18, rd21(encoding)
-    std::uint32_t encrypted{0};//encrypted (doc: 0=no, 1=yes, 2=unkn) on read: objects 0 and encrypted yes rd18
-    std::unordered_map<std::uint32_t, dwgPageInfo >pages;//index, size, offset
+    std::uint64_t compressed{1};//is compressed? 1=no, 2=yes rd18, rd21(encoding)
+    std::uint64_t encrypted{0};//encrypted (doc: 0=no, 1=yes, 2=unkn) on read: objects 0 and encrypted yes rd18
+    std::unordered_map<std::uint64_t, dwgPageInfo >pages;//index, size, offset
     std::uint64_t size{0};//size of section,  2000- rd15, rd18, rd21 (data size)
     std::uint64_t pageCount{0}; //number of pages (dwgPageInfo) in section rd18, rd21
     std::uint64_t maxSize{0}; //max decompressed size (needed??) rd18 rd21
