+++ src/intern/dwgutil.h
@@ -57,7 +57,7 @@ public:
     // Number of decompressed bytes produced by the last decompress18/21 call.
     // Callers of fixed-size pages (parseSysPage) require an exact-window fill;
     // data pages are input-bounded so a partial fill is normal.
-    std::uint32_t decompressedBytes() const { return decompPos; }
+    std::uint64_t decompressedBytes() const { return decompPos; }
 
 private:
     std::uint32_t litLength18();
@@ -67,13 +67,13 @@ private:
 
     std::uint32_t longCompressionOffset();
     std::uint32_t long20CompressionOffset();
-    std::uint32_t twoByteOffset(std::uint32_t *ll);
+    std::uint32_t twoByteOffset(std::uint64_t *ll);
 
     std::uint8_t compressedByte(void);
-    std::uint8_t compressedByte(const std::uint32_t index);
+    std::uint8_t compressedByte(const std::uint64_t index);
     std::uint32_t compressedHiByte(void);
     bool compressedInc(const std::int32_t inc = 1);
-    std::uint8_t decompByte(const std::uint32_t index);
+    std::uint8_t decompByte(const std::uint64_t index);
     void decompSet(const std::uint8_t value);
     bool buffersGood(void);
     void copyBlock21(const std::uint32_t length);
@@ -81,12 +81,12 @@ private:
     // Decode state — instance members (formerly static, which made decompress
     // non-reentrant and was a verbose/non-verbose nondeterminism hazard).
     std::uint8_t *compressedBuffer{nullptr};
-    std::uint32_t compressedSize{0};
-    std::uint32_t compressedPos{0};
+    std::uint64_t compressedSize{0};
+    std::uint64_t compressedPos{0};
     bool    compressedGood{true};
     std::uint8_t *decompBuffer{nullptr};
-    std::uint32_t decompSize{0};
-    std::uint32_t decompPos{0};
+    std::uint64_t decompSize{0};
+    std::uint64_t decompPos{0};
     bool    decompGood{true};
 
     static const std::uint8_t CopyOrder21_01[];
